Refactor link create/edit to class-based views
The views to create and edit classes were implemented as function-based
views. Class-based views are preferred over function-based, as they are 
the more modern django way of doing things. Plus, Because they are
implemented as classes, most boilerplate code and logic is already taken
care of for us, and where we implement complex logic more than once we 
can take advantage of inheritance, as we do here.

Also moves the template from being owned by the links app to being part
of the project, as it is not being implemented as a reusable app.