Acceptance tests suite & app & command
There's a new app in the repo called acceptance which contains a command
to
run the acceptance tests, all of which are inside the app's folder.

To run the one existing test use `python manage.py acceptance`

Splinter is the browser driver, and selenium is needed to provide the
phantomjs webdriver bridge.