Fixed up the footer to remove unnecessary elements
The crown copyright text remains but the image (which was there by the grace of
CSS) has been removed. The most appropriate way to remove the OGL licence seemed
to be to duplicate base.htm and remove the element entirely. This way we if ever
use the GOV.UK template as a submodule it won't clash. If only it were more
modular