Register usage when clicking through to link
Implementing basic usage logging when clicking through to an internal
link, or clicking the “yes” button on the interstitial for an external
link (viewing the interstitial and then going back does not count as
usage).

Adds a functional but unstyled “X times used today” to the detail page
for testing purposes.

The “yes” button on the interstitial changes to be a form button rather
than an anchor, so the styling there will need a bit of finessing.