Replaced options w radios. Moved errors, moved headers.
The select and option elements that previously allowed the user to select an
organisation for their team have been replaced by radio buttons, and the 'free text' option
looks a bit more like the radio buttons. The validation for preventing both from being
used still happens in the backend and is not yet as bold as it could be, but that will come
in another PR.

I also moved headers around and changed the containers for form elements to make them use
the GOV.UK stuff more, and made the tone a little more conversational, like the links form,
following gds guidelines.