Enable lighthouse to run in an internetless environment
In order to run in an environment where we can't pull down pip packages
we need to be able to install locally from a cached version. These
changes make it so that we always install from a cached version and if
we have internet, the default, then we can download the dependencies to
the cache. If we don't have internet, indicated to us by the environment
variable `internet_access` then we only install from the cache.