User favourites
Many to many relationship quite easily gives users favourites lists.

Now the UI Elements for Favourites...

Button: It posts to an endpoint called favourites which is attached to the user,
rather than the other way around. Buttons currently headed with a small
header explaining whether link is in favourite or not.

List: On the User Detail page, the user's favourite Tools are listed
beneath their most used tools.