Storing search queries whenever a user searches, with time, user, result length
This is a bit of funny one. SearchQuery and SearchTerm work fine, and in
order to store searches with their result length the search view had to
be overriden. However, whilst writing tests it appeared that the
haystack index was not beign reset before each test run and therefore
had the potential to bring up results which no longer existed in the DB.

As a result, the asserts which check for storing of search queries and
results were put into the most specific search tests in
`test_search_results.py` with `assertEquals` for the `result_length`
property, while the test which came up with more results than would've
come up in a production search was given a more flexible
`assertGreaterEqual` in order to make sure it still tested existence of
`results_length`. However, the other tests do test quite specifically.

WIP tests split up between being for the model and being for the web tests