API calls don’t need CSRF tokens
Because the initial API tests were naively written using the Django test
client, they were making requests in the context of a site user (so
had a session and a CSRF token). This wasn’t detecting that a real API
call would not be providing the token and so getting errors.

This updates the tests to use an “external” HTTP request against a live
server.

Also adds a test for getting a 404 on an invalid API link URL.