Display some instructive coloured text instead of a bell
Bell is a bit ambiguous, especially as it normally means notification of something happening.
Instead I added some concise text which says what the user needs to do and coloured it in an
attention-grabbing but inoffensive way. As well as that, I felt that a
good old-fashioned title on the <a> element would help provide some
context for powerusers.