CSV DTA, bigger lede, larger link, in grid.
A CTA is useful on the usage stats page because it's bigger and more beautiful.
The lede should be bigger especially since it is very little text. The whole page was out of the grid,
too, but it was in a grid row element so I just removed that element.