A blank search page is no longer super broken.
Because of the whole 'storing queries' thing an empty search page caused a 500, because
no search query was there to be stored. Also, when that problem *was* fixed it caused
another problem whereby the word 'None' appeared in the search box, since the value of
the context variable, 'q', was indeed None. So I fixed that too. There are also two new
tests to verify that this behaviour does not change, one for each of the cases that either
the querystring param q does not exist, or that it is empty. :thumbsup: