No longer showing the 20 most recent searches, CSV added
Since the 20 most recent searches are no longer being shown there's no need for
the search stats view to be a listview so it's a templateview now. Also, I added some
lede texts to that page explaining what it is plus a CSV link with explanation.
A CSV CTA is now also sitting beneath the table.