Removed the cookie banner and any code which relies on it.
There's some GOV.UK code which will handle the 'user has seen the cookie banner' thing,
which naturally will add a cookie. I left the code in the repo because it's vendor
code, but I removed it from our base.htm layout file because it wasn't being used. What
we've lost then is the ability to easily set a cookie on the front end, but that's OK and
if we need it later we can add it back in.