The 'next' parameter is now respected
The 'next' parameter in the URL (i.e. where the user intended to go when they
navigated to the site) will be respected. It will skip the onboarding
processs for new users. In my opinion, this is better than forcing them
to go through with it if they are trying to get to a particular page,
since whether they add more info or not is not of extreme importance
to the measuring of usage. They will still be shown the onboarding stuff
in other login circumstances.