Moved the .common testing functions into a single file.
For the existing files, I put a depreceation warning on each function, pointing to the new one in
case they are accidentally used. But each import should now point to the new single file.

This comes from noticing that some of them straddle lines between one domain and another, so
it makes sense to keep them in one place.

As for the depreceation warning, it had to have a type of 'Warning' rather than 'DepreceationWarning'
so that it would show up during testing.