Leave & Join buttons now POST, not GET
The leave team/join team buttons were actually just `<a>` links which
pointed to the leave and join links for that team. It is more
appropriate that they should be buttons whose parent form will POST to
the leave or join endpoint which should be an RPC-style endpoint because
GET requests are not meant to be destructive.

So that's what I've implemented. I had to modify the existing leave/join
tests to do it, and also put in some asserts to ensure that the old
leave/join buttons had been removed, too. They look exactly the same.
The only real change was to change on the backend was to change the
word `get` to `post`.

:snake: