Combined Link List view with Search view
Search now takes place on the Link List page.

How does it work? Link List is still a ListView, but it imports and uses
Haystack's SearchQuerySet and AutoQuery in order to do searches on the
various fields which make up a Link/Tool.

Name, categories, text (desc) and "externality" all are indexed. The
only one which features custom behaviour is "externality". In order to
make it text-searchable, i.e., that q=external would bring up links with
external checked as true, a custom Haystack `SearchField` based on
`CharField` was used, whose `convert` function was overridden to convert
the boolean into text.