Track lighthouse API usage in lighthouse
Creates a new default Link, for the API. When API posts are made,
update the usage of the API with activity from the owner of the app
involved in the API usage.

One test has to be run in isolation because of a Django bug around
database transactions, so the test runner script and README have been
updated to reflect this.