Update Tools as anybody
This widens the 'Edit' availability to any user whatsoever. Anybody can see the
Edit button and thus anybody can edit a link. However the most recent editor is the only
one shown alongside the owner on the Tool detail page. All the other editors are still stored
with the time they edited (so a list of edits is stored basically, without
any info about the edit itself) but are not shown anywhere yet.