Filter by favourites
There's now a checkbox on the tool list page for 'My favourites'
which, when checked, will be applied along with any other filters
currently applied.

In order to do this, the ID needs to be indexed (as "key") so that
haystack can filter on it, with an array of IDs which is built from the
user's favourites.