Markdown supported in tool descriptions.
This is done in order to allow the API tool to have a description with a link in it.

It had more work than expected, as the link list page required a special summarised
description for each link. These are formed by getting the first line, turning it into
html, stripping tags adn then trimming it to 140 chars or less with ... for
truncation.