Don't validate dest for Lighthouse/API
Both of the default links, Lighthouse and Lighthouse API, have 'invalid' URLs
according to django. This means that when the models are saved the validation blows
up the attempt. So a custom form class was introduced to make it possible to save
without this field preventing due to validation errors