Stop filter buttons flashing when JS is on
When JavaScript is available, the filter buttons are hidden so that merely
clicking the checkboxes will filter the page. Now, the filter buttons
will already be hidden if a context var says they should be.

This context var gets its information from a session variable that is
set when the user logs on. When the user logs in, some JS runs to make a
hidden input part of a form by giving it a 'name' attribute. This hidden
input, whose name is 'js', is used as part of the login POST and sets
the session variable to indicate whether the user has JS or not.

The result is the filter buttons being pre-hidden if the user definitely
has JS.