Prevent Sass compilation during request handling
This change fixes the issues we were seeing with CSS files 404ing. The
root cause was that the `COMPRESS_REBUILD_TIMEOUT` was too low (2 seconds),
meaning that between deployment and the uwsgi reload Compress believed
that it needed to rebuild the Sass file, despite no changes in the
content.

To fix this I have removed `COMPRESS_REBUILD_TIMEOUT`, letting it's
default of 30 days (so long as the content doesn't change) remain, and
using the `COMPRESS_OFFLINE` to prevent compress from trying to do a
recompile during handling the requests.